pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skills-verified"
version = "0.1.0"
description = "AI Agent Trust Scanner — CVE, vulnerability, and guardrails analysis"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "gitpython>=3.1",
    "Levenshtein>=0.25",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = ["openai>=1.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
skills-verified = "skills_verified.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
src = ["src"]
